requirements.txt
# Core dependencies
pydantic>=2.0.0
fastapi>=0.100.0
uvicorn>=0.20.0

# Data Discovery dependencies
aiohttp>=3.8.0
cachetools>=5.0.0
ratelimit>=2.2.1
httpx>=0.24.0
asyncpg>=0.27.0

# Storage
python-dotenv>=1.0.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.20.0
pytest-cov>=4.0.0
aioresponses>=0.7.4
ratelimit>=2.2.1

# Other dependencies
flake8>=6.0.0
black>=23.0.0
isort>=5.0.0
openai>=1.12.0
tenacity>=8.2.3